# perf: make Mul.is_zero/is_infinite faster