#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->

I noticed this was slow in https://github.com/sympy/sympy/issues/24480#issuecomment-1375538156.

#### Brief description of what is fixed or changed

Rewrite `Mul._eval_is_zero` and `Mul._eval_is_infinite` to reuse the same helper function and made that helper function faster for the common case where assumptions queries give None. Now in a large Mul like `Mul(*symbols('x:1000'))` a query like is_zero is around 10x faster. Still it is unnecessarily slow because of _eval_is_even and _eval_is_odd using fraction though.

On master:
```python
In [1]: m = Mul(*symbols('x:1000'))

In [2]: %time m.is_zero
CPU times: user 372 ms, sys: 4 ms, total: 376 ms
Wall time: 377 ms

In [1]: m = Mul(*symbols('x:3'))

In [2]: %time m.is_zero
CPU times: user 8 ms, sys: 0 ns, total: 8 ms
Wall time: 7.34 ms
```
With PR:
```python
In [1]: m = Mul(*symbols('x:1000')) # big speedup for big Muls

In [2]: %time m.is_zero
CPU times: user 32 ms, sys: 0 ns, total: 32 ms
Wall time: 30.5 ms

In [1]: m = Mul(*symbols('x:3')) # no slowdown for small Muls

In [2]: %time m.is_zero
CPU times: user 4 ms, sys: 0 ns, total: 4 ms
Wall time: 5.92 ms
```
(Note if you want to time this then be careful about the fact that assumptions queries are cached - you need to restart Python each time to run timings. Also don't use timeit)

#### Other comments

Profiling this operation under the PR gives:
```python
In [1]: m = Mul(*symbols('x:1000'))

In [2]: %prun -s cumulative m.is_positive

         74941 function calls (68885 primitive calls) in 0.084 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.000    0.000    0.084    0.084 {built-in method builtins.exec}
        1    0.000    0.000    0.084    0.084 <string>:1(<module>)
   3041/1    0.003    0.000    0.084    0.084 assumptions.py:498(getit)
     17/1    0.004    0.000    0.084    0.084 assumptions.py:510(_ask)
        1    0.000    0.000    0.073    0.073 mul.py:1652(_eval_is_even)
        1    0.003    0.003    0.073    0.073 radsimp.py:1053(fraction)
   2002/2    0.003    0.000    0.061    0.031 cache.py:69(wrapper)
        2    0.001    0.001    0.061    0.030 operations.py:52(__new__)
        1    0.006    0.006    0.052    0.052 mul.py:197(flatten)
        2    0.009    0.004    0.022    0.011 mul.py:467(_gather)
        1    0.000    0.000    0.021    0.021 mul.py:33(_mulsort)
        1    0.001    0.001    0.021    0.021 {method 'sort' of 'list' objects}
      999    0.006    0.000    0.020    0.000 basic.py:205(compare)
12007/11007    0.005    0.000    0.011    0.000 {built-in method builtins.isinstance}
     1998    0.004    0.000    0.010    0.000 symbol.py:340(_hashable_content)
     1000    0.006    0.000    0.006    0.000 exponential.py:211(__instancecheck__)
...
```
This shows that the almost all the remaining time is spent in `_eval_is_even` here:
https://github.com/sympy/sympy/blob/6f0fd470a4f7a582c21d3a610798943817d0a334/sympy/core/mul.py#L1590-L1601
The `_eval_is_even` handler should be changed and simplified. Functions like `fraction` that create new expressions should not be used during assumptions queries. A simple implementation of `_eval_is_even` should be able to exit after just looking at a single arg of the Mul. What this shows is why assumptions handlers should not be made too clever. It is important to understand that an assumptions handler like `_eval_is_even` will be called even when the assumptions query is `is_zero` because ofo the inference rules. So any extra processing for the relatively unimportant `is_even` case has the potential to slow down any operation involving large expressions.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* core
  * Assumptions queries on large Muls were made more efficient e.g. given `m = Mul(*symbols('x:1000'))` a first call to `m.is_zero` is around 10x faster. Assumptions queries often happen implicitly so this can have a broad effect on the speed of operations with large expressions.
<!-- END RELEASE NOTES -->