@smichr 2023-01-09
> Functions like `fraction` that create new expressions should not be used during assumptions queries. A simple implementation of `_eval_is_even` should be able to exit after just looking at a single arg of the Mul.

You've emphasized this repeatedly (and I appreciate that). But `fraction` doesn't create new expressions as I understand. Per docstring: "This function will not make any attempt to simplify nested fractions or to do any term rewriting at all." So it seems to me that looking at the numerator and denominator is not different than looking at individual args.

> exit after looking at a single arg of the Mul

I surely am misunderstanding what you mean because `Dummy(odd=1)*Dummy(even=1)` cannot exit after looking at the first arg.