@oscarbenjamin 2023-01-10
> But `fraction` doesn't create new expressions as I understand.

It creates two new Muls. In the example `fraction(Mul(*symbols('x:1000')))` one of those Muls is the same as the input expression so it's rebuilding the same Mul. This also leads to `Mul.flatten` being called which does many more assumptions queries.

> I surely am misunderstanding what you mean because `(Dummy(odd=1)*Dummy(even=1)).is_even` cannot exit after looking at the first arg.

Symbols with even/odd assumptions are very rarely used. The common case is that there is exactly one Number (usually a rational coefficient) and every other arg in the Mul will have integer=None. We only need one in arg in the Mul to have e.g. even=None and then the result is just None. In the case of `Mul(*symbols('x:1000'))` we can just look at the first arg and see that it has no relevant assumptions set and return None. Calling fraction instead loops through the every arg and rebuilds the Mul with no possibility to shortcut.