@smichr 2023-01-10
> Calling fraction instead loops through the every arg and rebuilds the Mul with no possibility to shortcut.

We can short circuit on the first *None arg* -- on that I agree. Dealing with `fraction` only makes that become the first *two* None args, one from numerator and one from denominator. One only loops through every arg if every arg has some assumption of relevance. But that is not a change under consideration here so it's best left to further discussion when changes to `is_even` are proposed.